+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ---
 
+# [v0.5.0](https://github.com/DaemonDude23/container-image-replicator/releases/tag/v0.5.0) - Jan 23 2023
+
+**Enhancements**
+
+- Add CLI flag `--force-pull-push` to force pulling/pushing images even if the tag exists in the remote repository. Use `[]source.forcePull` and or `[]source.forcePush` to fine-tune these properties on a per-image basis instead of globally with this CLI flag.
+  - **Beware as this can count against rate limits. Use wisely!**
+
+**Housekeeping**
+
+- `types-PyYAML` package updated.
+- Added to and updated docs.
+
 # [v0.4.2](https://github.com/DaemonDude23/container-image-replicator/releases/tag/v0.4.2) - Jan 9 2023
 
 **Bugfixes**